tests: Move MessageAccessTests to test_message_flags.py.
